feat: 상대팀별 승률 조회에서 상대 전적 연동 추가

+++ android/app/src/main/java/com/yagubogu/presentation/stats/detail/VsTeamStatItem.kt
@@ -7,9 +7,9 @@ data class VsTeamStatItem(
     val rank: Int,
     val team: Team,
     val teamName: String,
-    val winCounts: Int,
-    val drawCounts: Int,
-    val loseCounts: Int,
+    val winCounts: Long,
+    val drawCounts: Long,
+    val loseCounts: Long,
     val winningPercentage: Double,
 ) {
     val teamEmoji: String = team.getEmoji()
